+++ amigo-lib/src/main/java/me/ele/amigo/Amigo.java
@@ -1,6 +1,7 @@
 package me.ele.amigo;
 
 import android.app.Application;
+import android.app.Instrumentation;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.content.pm.ApplicationInfo;
@@ -8,27 +9,28 @@ import android.content.pm.Signature;
 import android.content.res.AssetManager;
 import android.content.res.Resources;
 import android.os.Build;
+import android.os.Handler;
 import android.os.Process;
 import android.text.TextUtils;
 import android.util.ArrayMap;
 import android.util.Log;
 
 import java.io.File;
-import java.io.IOException;
+import java.io.FileFilter;
 import java.lang.ref.WeakReference;
-import java.lang.reflect.Array;
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.security.NoSuchAlgorithmException;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
+import me.ele.amigo.hook.HookFactory;
+import me.ele.amigo.reflect.MethodUtils;
 import me.ele.amigo.release.ApkReleaser;
 import me.ele.amigo.utils.CommonUtils;
+import me.ele.amigo.utils.CrcUtils;
+import me.ele.amigo.utils.PermissionChecker;
 import me.ele.amigo.utils.ProcessUtils;
+import me.ele.amigo.utils.component.ReceiverFinder;
 
 import static android.content.pm.PackageManager.GET_META_DATA;
 import static me.ele.amigo.compat.ActivityThreadCompat.instance;
@@ -36,14 +38,11 @@ import static me.ele.amigo.reflect.FieldUtils.getField;
 import static me.ele.amigo.reflect.FieldUtils.readField;
 import static me.ele.amigo.reflect.FieldUtils.readStaticField;
 import static me.ele.amigo.reflect.FieldUtils.writeField;
-import static me.ele.amigo.reflect.MethodUtils.getDeclaredMethod;
+import static me.ele.amigo.reflect.MethodUtils.getMatchedMethod;
 import static me.ele.amigo.reflect.MethodUtils.invokeMethod;
 import static me.ele.amigo.reflect.MethodUtils.invokeStaticMethod;
+import static me.ele.amigo.utils.ClassLoaderUtils.getRootClassLoader;
 import static me.ele.amigo.utils.CrcUtils.getCrc;
-import static me.ele.amigo.utils.DexUtils.getElementWithDex;
-import static me.ele.amigo.utils.DexUtils.getPathList;
-import static me.ele.amigo.utils.DexUtils.getRootClassLoader;
-import static me.ele.amigo.utils.DexUtils.injectSoAtFirst;
 import static me.ele.amigo.utils.FileUtils.copyFile;
 import static me.ele.amigo.utils.FileUtils.removeFile;
 
@@ -51,15 +50,18 @@ public class Amigo extends Application {
     private static final String TAG = Amigo.class.getSimpleName();
 
     public static final String SP_NAME = "Amigo";
-    public static final String NEW_APK_SIG = "new_apk_sig";
+    public static final String WORKING_PATCH_APK_CHECKSUM = "working_patch_apk_checksum";
     public static final String VERSION_CODE = "version_code";
 
-    private ClassLoader originalClassLoader;
-    private ClassLoader patchedClassLoader;
+    private static LoadPatchError loadPatchError;
+
     private SharedPreferences sharedPref;
 
+    private ClassLoader originalClassLoader;
+    private AmigoClassLoader patchedClassLoader;
+
     private AmigoDirs amigoDirs;
-    private PatchApk patchApk;
+    private PatchApks patchApks;
 
     @Override
     public void onCreate() {
@@ -67,38 +69,50 @@ public class Amigo extends Application {
         try {
             originalClassLoader = getClassLoader();
             sharedPref = getSharedPreferences(SP_NAME, MODE_MULTI_PROCESS);
+            amigoDirs = AmigoDirs.getInstance(this);
+            patchApks = PatchApks.getInstance(this);
+            String workingPatchApkChecksum = sharedPref.getString(WORKING_PATCH_APK_CHECKSUM, "");
             try {
-                ensureAmigoDirs();
+                Log.e(TAG, "working checksum: " + workingPatchApkChecksum);
                 if (checkUpgrade()) {
                     throw new RuntimeException("Host app has upgrade");
                 }
-                if (!patchApk.exists()) {
+                if (TextUtils.isEmpty(workingPatchApkChecksum)
+                        || !patchApks.exists(workingPatchApkChecksum)) {
                     throw new RuntimeException("Patch apk doesn't exists");
                 }
-                if (!checkSignature()) {
+                if (!checkSignature(workingPatchApkChecksum)) {
+                    loadPatchError = LoadPatchError.record(LoadPatchError.SIG_ERR, null);
                     throw new RuntimeException("Patch apk has illegal signature");
                 }
-                if (!checkVersion()) {
-                    throw new RuntimeException("Patch apk version can't downgrade");
-                }
-
             } catch (RuntimeException e) {
                 e.printStackTrace();
-                clear(this);
+                if (ProcessUtils.isMainProcess(this)) {
+                    // clear is a dangerous operation, only need to be operated by main process
+                    doClear(this);
+                }
                 runOriginalApplication();
                 return;
             }
 
-            if (!ProcessUtils.isMainProcess(this) && isPatchApkFirstRun()) {
+            // ensure load dex process always run host apk not patch apk
+            if (ProcessUtils.isLoadDexProcess(this)) {
+                Log.e(TAG, "load dex process");
+                runOriginalApplication();
+                return;
+            }
+
+            if (!ProcessUtils.isMainProcess(this) && isPatchApkFirstRun(workingPatchApkChecksum)) {
                 Log.e(TAG, "None main process and patch apk is not released yet");
                 runOriginalApplication();
                 return;
             }
 
             // only release loaded apk in the main process
-            runPatchApk();
+            runPatchApk(workingPatchApkChecksum);
         } catch (LoadPatchApkException e) {
             e.printStackTrace();
+            loadPatchError = LoadPatchError.record(LoadPatchError.LOAD_ERR, e);
             try {
                 runOriginalApplication();
             } catch (Throwable e2) {
@@ -109,38 +123,94 @@ public class Amigo extends Application {
         }
     }
 
-    private void runPatchApk() throws LoadPatchApkException {
+    private void runPatchApk(String checksum) throws LoadPatchApkException {
         try {
-            Log.e(TAG, "patchApkChecksum-->" + patchApk.checksum() + ", sig--->" + sharedPref.getString(NEW_APK_SIG, ""));
-            if (isPatchApkFirstRun() || !isOptedDexExists()) {
+            Log.e(TAG, "patchApkChecksum-->" + checksum + ", sp record checksum--->"
+                    + sharedPref.getString(WORKING_PATCH_APK_CHECKSUM, ""));
+            if (isPatchApkFirstRun(checksum) || !isOptedDexExists(checksum)) {
                 // TODO This is workaround for now, refactor in future.
-                sharedPref.edit().remove(patchApk.checksum()).commit();
-                releasePatchApk();
+                sharedPref.edit().remove(checksum).commit();
+                releasePatchApk(checksum);
             } else {
-                checkDexAndSoChecksum();
+                checkDexAndSoChecksum(checksum);
             }
 
-            AmigoClassLoader amigoClassLoader = new AmigoClassLoader(patchApk.patchPath(), getRootClassLoader());
+            String dexPathes = getDexPath(checksum);
+            AmigoClassLoader amigoClassLoader = new AmigoClassLoader(dexPathes,
+                    AmigoDirs.getInstance(this).dexOptDir(checksum),
+                    AmigoDirs.getInstance(this).libDir(checksum).getAbsolutePath(),
+                    getRootClassLoader());
             setAPKClassLoader(amigoClassLoader);
-            setDexElements(amigoClassLoader);
-            setNativeLibraryDirectories(amigoClassLoader);
+            patchedClassLoader = amigoClassLoader;
 
             AssetManager assetManager = AssetManager.class.newInstance();
-            Method addAssetPath = getDeclaredMethod(AssetManager.class, "addAssetPath", String.class);
+            Method addAssetPath = getMatchedMethod(AssetManager.class, "addAssetPath", String.class);
             addAssetPath.setAccessible(true);
-            addAssetPath.invoke(assetManager, patchApk.patchPath());
+            addAssetPath.invoke(assetManager, patchApks.patchPath(checksum));
             setAPKResources(assetManager);
 
-            patchedClassLoader = amigoClassLoader;
+            setApkInstrumentation();
+            setApkHandler();
+            dynamicRegisterReceivers(amigoClassLoader);
+
+            sharedPref.edit().putString(WORKING_PATCH_APK_CHECKSUM, checksum).commit();
+            clearOldPatches(checksum);
+            installHook(amigoClassLoader);
             runPatchedApplication();
         } catch (Exception e) {
             throw new LoadPatchApkException(e);
         }
     }
 
-    private void releasePatchApk() throws Exception {
+  private String getDexPath(String checksum) throws LoadPatchApkException {
+        File[] patchDexFiles = AmigoDirs.getInstance(this).dexDir(checksum).listFiles(
+                new FileFilter() {
+                    @Override
+                    public boolean accept(File pathname) {
+                        return pathname.getName().endsWith(".dex");
+                    }
+                });
+        String dexPath = "";
+        if (patchDexFiles != null && patchDexFiles.length > 0) {
+            for (File patchDex : patchDexFiles) {
+                dexPath += ":" + patchDex.getAbsolutePath();
+            }
+        } else {
+            LoadPatchApkException e= new LoadPatchApkException("Amigo: no dexes avilable");
+            e.fillInStackTrace();
+            throw  e;
+        }
+        return dexPath;
+    }
+
+    private void installHook(AmigoClassLoader amigoClassLoader) throws Exception {
+        Class hookFactoryClazz = amigoClassLoader.loadClass(HookFactory.class.getName());
+        MethodUtils.invokeStaticMethod(hookFactoryClazz, "install", this, amigoClassLoader);
+    }
+
+    private void dynamicRegisterReceivers(ClassLoader classLoader) {
+        ReceiverFinder.registerNewReceivers(this, classLoader);
+    }
+
+    private void setApkInstrumentation() throws Exception {
+        Instrumentation oldInstrumentation = (Instrumentation) readField(instance(), "mInstrumentation", true);
+        Log.e(TAG, "oldInstrumentation--->" + oldInstrumentation);
+        AmigoInstrumentation instrumentation = new AmigoInstrumentation(oldInstrumentation);
+        writeField(instance(), "mInstrumentation", instrumentation, true);
+        Log.e(TAG, "setApkInstrumentation success classloader-->" + instrumentation.getClass().getClassLoader());
+    }
+
+    private void setApkHandler() throws Exception {
+        Handler handler = (Handler) readField(instance(), "mH", true);
+        Object callback = readField(handler, "mCallback", true);
+        AmigoCallback value = new AmigoCallback(this, patchedClassLoader, (Handler.Callback) callback);
+        writeField(handler, "mCallback", value);
+        Log.e(TAG, "hook handler success");
+    }
+
+    private void releasePatchApk(String checksum) throws Exception {
         //clear previous working dir
-        clearWithoutPatch(this);
+        clearWithoutPatchApk(checksum);
 
         //start a new process to handle time-tense operation
         ApplicationInfo appInfo = getPackageManager().getApplicationInfo(getPackageName(), GET_META_DATA);
@@ -157,25 +227,19 @@ public class Amigo extends Application {
         Log.e(TAG, String.format("layoutName-->%s, themeName-->%s", layoutName, themeName));
         Log.e(TAG, String.format("layoutId-->%d, themeId-->%d", layoutId, themeId));
 
-        ApkReleaser.getInstance(this).work(layoutId, themeId);
+        ApkReleaser.getInstance(this).work(checksum, layoutId, themeId);
         Log.e(TAG, "release apk once");
     }
 
-    private boolean isPatchApkFirstRun() {
-        return !sharedPref.getString(NEW_APK_SIG, "").equals(patchApk.checksum());
+    private boolean isPatchApkFirstRun(String checksum) {
+        return !sharedPref.getString(WORKING_PATCH_APK_CHECKSUM, "").equals(checksum);
     }
 
-    private boolean isOptedDexExists() {
-        if (amigoDirs.dexOptDir().listFiles() != null) {
-            return amigoDirs.dexOptDir().listFiles().length > 0;
+    private boolean isOptedDexExists(String checksum) {
+        if (amigoDirs.dexOptDir(checksum).listFiles() != null) {
+            return amigoDirs.dexOptDir(checksum).listFiles().length > 0;
         }
         return false;
-    }
-
-    private void ensureAmigoDirs() throws Exception {
-        AmigoDirs.init(this);
-        amigoDirs = AmigoDirs.getInstance();
-        patchApk = PatchApk.getInstance();
     }
 
     private boolean checkUpgrade() {
@@ -195,7 +259,7 @@ public class Amigo extends Application {
         String applicationName = (String) readStaticField(acd, "n");
         Application application =
                 (Application) originalClassLoader.loadClass(applicationName).newInstance();
-        Method attach = getDeclaredMethod(Application.class, "attach", Context.class);
+        Method attach = getMatchedMethod(Application.class, "attach", Context.class);
         attach.setAccessible(true);
         attach.invoke(application, getBaseContext());
         setAPKApplication(application);
@@ -206,18 +270,17 @@ public class Amigo extends Application {
         setAPKClassLoader(patchedClassLoader);
         Class acd = patchedClassLoader.loadClass("me.ele.amigo.acd");
         String applicationName = (String) readStaticField(acd, "n");
-        Application application =
-                (Application) patchedClassLoader.loadClass(applicationName).newInstance();
-        Method attach = getDeclaredMethod(Application.class, "attach", Context.class);
+        Application application = (Application) patchedClassLoader.loadClass(applicationName).newInstance();
+        Method attach = getMatchedMethod(Application.class, "attach", Context.class);
         attach.setAccessible(true);
         attach.invoke(application, getBaseContext());
         setAPKApplication(application);
         application.onCreate();
     }
 
-    private void checkDexAndSoChecksum() throws IOException, NoSuchAlgorithmException {
+    private void checkDexAndSoChecksum(String apkChecksum) throws Exception {
         SharedPreferences sp = getSharedPreferences(SP_NAME, MODE_MULTI_PROCESS);
-        File[] dexFiles = amigoDirs.dexDir().listFiles();
+        File[] dexFiles = amigoDirs.dexDir(apkChecksum).listFiles();
         for (File dexFile : dexFiles) {
             String savedChecksum = sp.getString(dexFile.getAbsolutePath(), "");
             String checksum = getCrc(dexFile);
@@ -226,7 +289,7 @@ public class Amigo extends Application {
             }
         }
 
-        File[] dexOptFiles = amigoDirs.dexOptDir().listFiles();
+        File[] dexOptFiles = amigoDirs.dexOptDir(apkChecksum).listFiles();
         for (File dexOptFile : dexOptFiles) {
             String savedChecksum = sp.getString(dexOptFile.getAbsolutePath(), "");
             String checksum = getCrc(dexOptFile);
@@ -237,7 +300,7 @@ public class Amigo extends Application {
             }
         }
 
-        File[] nativeFiles = amigoDirs.libDir().listFiles();
+        File[] nativeFiles = amigoDirs.libDir(apkChecksum).listFiles();
         if (nativeFiles != null && nativeFiles.length > 0) {
             for (File nativeFile : nativeFiles) {
                 String savedChecksum = sp.getString(nativeFile.getAbsolutePath(), "");
@@ -249,21 +312,7 @@ public class Amigo extends Application {
         }
     }
 
-    private void setNativeLibraryDirectories(AmigoClassLoader hackClassLoader)
-            throws InvocationTargetException, NoSuchMethodException, InstantiationException, IllegalAccessException, NoSuchFieldException {
-        File libDir = amigoDirs.libDir();
-        injectSoAtFirst(hackClassLoader, libDir.getAbsolutePath());
-        libDir.setReadOnly();
-        File[] libs = libDir.listFiles();
-        if (libs != null && libs.length > 0) {
-            for (File lib : libs) {
-                lib.setReadOnly();
-            }
-        }
-    }
-
-    private void setAPKResources(AssetManager newAssetManager)
-            throws NoSuchMethodException, IllegalAccessException, InvocationTargetException, ClassNotFoundException {
+    private void setAPKResources(AssetManager newAssetManager) throws Exception {
         invokeMethod(newAssetManager, "ensureStringBlocks");
 
         Collection<WeakReference<Resources>> references;
@@ -311,40 +360,16 @@ public class Amigo extends Application {
         }
     }
 
-    private void setAPKClassLoader(ClassLoader classLoader)
-            throws IllegalAccessException, NoSuchMethodException, ClassNotFoundException, InvocationTargetException {
+    private void setAPKClassLoader(ClassLoader classLoader) throws Exception {
         writeField(getLoadedApk(), "mClassLoader", classLoader);
     }
 
-    private void setDexElements(ClassLoader classLoader) throws NoSuchFieldException, IllegalAccessException {
-        Object dexPathList = getPathList(classLoader);
-        File[] listFiles = amigoDirs.dexDir().listFiles();
-
-        List<File> validDexes = new ArrayList<>();
-        for (File listFile : listFiles) {
-            if (listFile.getName().endsWith(".dex")) {
-                validDexes.add(listFile);
-            }
-        }
-        File[] dexes = validDexes.toArray(new File[validDexes.size()]);
-        Object originDexElements = readField(dexPathList, "dexElements");
-        Class<?> localClass = originDexElements.getClass().getComponentType();
-        int length = dexes.length;
-        Object dexElements = Array.newInstance(localClass, length);
-        for (int k = 0; k < length; k++) {
-            Array.set(dexElements, k, getElementWithDex(dexes[k], amigoDirs.dexOptDir()));
-        }
-        writeField(dexPathList, "dexElements", dexElements);
-    }
-
-    private void setAPKApplication(Application application)
-            throws InvocationTargetException, NoSuchMethodException, ClassNotFoundException, IllegalAccessException {
+    private void setAPKApplication(Application application) throws Exception {
         Object apk = getLoadedApk();
         writeField(apk, "mApplication", application);
     }
 
-    private static Object getLoadedApk()
-            throws IllegalAccessException, NoSuchMethodException, InvocationTargetException, ClassNotFoundException {
+    private static Object getLoadedApk() throws Exception {
         Map<String, WeakReference<Object>> mPackages = (Map<String, WeakReference<Object>>) readField(instance(), "mPackages", true);
         for (String s : mPackages.keySet()) {
             WeakReference wr = mPackages.get(s);
@@ -355,42 +380,33 @@ public class Amigo extends Application {
         return null;
     }
 
-    public static void workLater(Context context) {
-        workLater(context, PatchApk.getInstance().patchFile());
-    }
-
     public static void workLater(Context context, File patchFile) {
-        if (context == null) {
-            throw new NullPointerException("param context cannot be null");
+        checkPatchApk(context, patchFile);
+        String patchChecksum = CrcUtils.getCrc(patchFile);
+        if (!PatchApks.getInstance(context).exists(patchChecksum)) {
+            copyFile(patchFile, PatchApks.getInstance(context).patchFile(patchChecksum));
         }
 
-        if (patchFile == null) {
-            throw new NullPointerException("param patchFile cannot be null");
-        }
-
-        if (!patchFile.exists()) {
-            throw new IllegalArgumentException("param patchFile doesn't exist");
-        }
-
-        if (!patchFile.canRead()) {
-            throw new IllegalArgumentException("param patchFile cannot be read");
-        }
-        if (!patchFile.getAbsolutePath().equals(PatchApk.getInstance().patchPath())) {
-            copyFile(patchFile, PatchApk.getInstance().patchFile());
-        }
-
-        AmigoService.start(context, true);
-    }
-
-    public static void work(Context context) {
-        work(context, PatchApk.getInstance().patchFile());
+        AmigoService.start(context, patchChecksum, true);
     }
 
     public static void work(Context context, File patchFile) {
-        if (context == null) {
-            throw new NullPointerException("param context cannot be null");
+        checkPatchApk(context, patchFile);
+        String patchChecksum = CrcUtils.getCrc(patchFile);
+        if (!PatchApks.getInstance(context).exists(patchChecksum)) {
+            copyFile(patchFile, PatchApks.getInstance(context).patchFile(patchChecksum));
         }
 
+        context.getSharedPreferences(SP_NAME, MODE_MULTI_PROCESS)
+                .edit()
+                .putString(Amigo.WORKING_PATCH_APK_CHECKSUM, patchChecksum)
+                .commit();
+        AmigoService.start(context, patchChecksum, false);
+        System.exit(0);
+        Process.killProcess(Process.myPid());
+    }
+
+    private static void checkPatchApk(Context context, File patchFile) {
         if (patchFile == null) {
             throw new NullPointerException("param apkFile cannot be null");
         }
@@ -403,19 +419,15 @@ public class Amigo extends Application {
             throw new IllegalArgumentException("param apkFile cannot be read");
         }
 
-        if (!patchFile.getAbsolutePath().equals(PatchApk.getInstance().patchPath())) {
-            copyFile(patchFile, PatchApk.getInstance().patchFile());
+        if (!PermissionChecker.checkPatchPermission(context, patchFile)) {
+            throw new IllegalStateException("patch apk cannot request more permissions than host");
         }
-
-        AmigoService.start(context, false);
-        System.exit(0);
-        Process.killProcess(Process.myPid());
     }
 
-    private boolean checkSignature() {
+    private boolean checkSignature(String checksum) {
         try {
             Signature appSig = CommonUtils.getSignature(this);
-            Signature patchSig = patchApk.signature(this);
+            Signature patchSig = patchApks.signature(checksum);
             return appSig.hashCode() == patchSig.hashCode();
         } catch (Exception e) {
             e.printStackTrace();
@@ -423,16 +435,9 @@ public class Amigo extends Application {
         return false;
     }
 
-    private boolean checkVersion() {
-        return patchApk.versionCode(this) >= CommonUtils.getVersionCode(this);
-    }
-
-    public static void clear(Context context) {
-        if (context == null) {
-            throw new NullPointerException("param context cannot be null");
-        }
-
-        removeFile(AmigoDirs.getInstance().amigoDir());
+    private static void doClear(Context context) {
+        Log.e(TAG, "clear");
+        removeFile(AmigoDirs.getInstance(context).amigoDir());
         context.getSharedPreferences(SP_NAME, MODE_MULTI_PROCESS)
                 .edit()
                 .clear()
@@ -440,17 +445,33 @@ public class Amigo extends Application {
                 .commit();
     }
 
-    private void clearWithoutPatch(Context context) {
-        if (context == null) {
-            throw new NullPointerException("param context cannot be null");
-        }
-
-        File[] files = amigoDirs.amigoDir().listFiles();
-        if (files != null && files.length > 0) {
-            for (File file : files) {
-                if (!file.getAbsolutePath().equals(patchApk.patchPath())) {
-                    removeFile(file, false);
+    private void clearWithoutPatchApk(String checksum) {
+        Log.e(TAG, "clear without patch");
+        File[] patchDirs = amigoDirs.amigoDir().listFiles();
+        if (patchDirs != null && patchDirs.length > 0) {
+            for (File patchDir : patchDirs) {
+                if (patchDir.getName().equals(checksum)) {
+                    File[] files = patchDir.listFiles();
+                    for (File file : files) {
+                        if (!file.getAbsolutePath().equals(patchApks.patchPath(checksum))) {
+                            Log.e(TAG, "remove file: " + file.getAbsolutePath());
+                            removeFile(file, false);
+                        }
+                    }
                 }
+            }
+        }
+    }
+
+    private void clearOldPatches(String exclude) {
+        Log.e(TAG, "clear old patches");
+        File[] patchDirs = amigoDirs.amigoDir().listFiles();
+        if (patchDirs == null || patchDirs.length == 0) {
+            return;
+        }
+        for (File patchDir : patchDirs) {
+            if (!patchDir.getName().equals(exclude)) {
+                removeFile(patchDir, true);
             }
         }
     }
@@ -465,9 +486,34 @@ public class Amigo extends Application {
         return classLoader != null && classLoader.getClass().getName().equals(AmigoClassLoader.class.getName());
     }
 
+    public static int workingPatchVersion(Context ctx) {
+        if (!hasWorked() || TextUtils.isEmpty(getWorkingPatchApkChecksum(ctx))) return -1;
+        return CommonUtils.getVersionCode(ctx, PatchApks.getInstance(ctx).patchFile(getWorkingPatchApkChecksum(ctx)));
+    }
+
+    public static String getWorkingPatchApkChecksum(Context ctx) {
+        if (!hasWorked()) return "";
+        return ctx.getSharedPreferences(SP_NAME, MODE_MULTI_PROCESS).getString(WORKING_PATCH_APK_CHECKSUM, "");
+    }
+
+    public static void clear(Context context) {
+        context.getSharedPreferences(SP_NAME, MODE_MULTI_PROCESS)
+                .edit()
+                .putString(WORKING_PATCH_APK_CHECKSUM, "")
+                .commit();
+    }
+
+    public static LoadPatchError getLoadPatchError() {
+        return loadPatchError;
+    }
+
     private static class LoadPatchApkException extends Exception {
         public LoadPatchApkException(Throwable throwable) {
             super(throwable);
+        }
+
+        public LoadPatchApkException(String msg) {
+
         }
     }
 }