+++ app/src/main/java/me/ele/app/amigo/activity/PatchedSingleTopActivity.java
@@ -1,0 +1,28 @@
+package me.ele.app.amigo.activity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
+
+import me.ele.app.amigo.R;
+
+public class PatchedSingleTopActivity extends AppCompatActivity {
+
+    public static final String TAG = PatchedSingleTopActivity.class.getSimpleName();
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        setContentView(R.layout.activity_second);
+        Log.d(TAG, "onCreate: " + hashCode());
+    }
+
+    @Override
+    protected void onNewIntent(Intent intent) {
+        super.onNewIntent(intent);
+        Log.d(TAG, "onNewIntent: " + intent.getStringExtra("extra"));
+    }
+}