+++ CHANGELOG.md
@@ -1,5 +1,41 @@
-## Version [0.1.0](https://github.com/eleme/Amigo/releases/tag/v0.1.0)
+## Version [0.4.2](https://github.com/eleme/Amigo/releases/tag/v0.4.2)
+_2016-10-15_
+* fix load receiver error
+* avoid service from mixed up
+
+## Version [0.4.1](https://github.com/eleme/Amigo/releases/tag/v0.4.1)
+_2016-10-11_
+* fix check permission bug
+
+## Version [0.4.0](https://github.com/eleme/Amigo/releases/tag/v0.4.0) (Deprecated)
+
+_2016-10-10_
+
+* support new added `service`
+* refactor ClassLoader replacement
+* support new added permission's check
+
+## Version [0.3.1](https://github.com/eleme/Amigo/releases/tag/v0.3.1)
+
+_2016-09-29_
+
+* fix Intent unmarshalling Exception [issue99](https://github.com/eleme/Amigo/issues/99)
+
+## Version [0.3.0](https://github.com/eleme/Amigo/releases/tag/v0.3.0)
+
+_2016-09-26_
+
+* support new added `activity` & `receiver`
+
+## Version [0.2.0](https://github.com/eleme/Amigo/releases/tag/v0.2.0)
+
+_2016-9-23_
+
+* Refactor the demo/develop app
+* Fix [re-apply patch apk bug](https://github.com/eleme/Amigo/issues/73)
+
+## Version [0.1.0](https://github.com/eleme/Amigo/releases/tag/v0.1.0) (Deprecated)
 
 _2016-09-14_
 
-first stable version
+* First stable version